Verificacao publisher controller

+++ API_Products/controllers/publisher_controller.py
@@ -23,7 +23,10 @@ def read_all_publishers() -> dict:
 
 def delete_publisher(dict_values: dict) -> dict:
     try:
-        publishers = publisher_db.delete_publishers_db(dict_values)
+        if not publisher_db.exists_publisher(dict_values):
+            publishers = publisher_db.delete_publishers_db(dict_values)
+        else:
+            raise Exception("Essa catogoria já existe e está sendo usada!")
         return dict(status=200, text=publishers)
     except Exception as error:        
         return dict(status=400, text=error.args[0])
